requirements/test.txt
parameterized==0.9.0
flake8==7.0.0
flake8-import-order==0.18.2
flake8-quotes==3.4.0
flake8-commas==2.1.0
pep8-naming==0.13.3
flake8-variables-names==0.0.6
django-debug-toolbar==4.3.0
